Add degree circle plot examples. Remove con_circle.py and include funcs into con_viz.py.
