+++ jumeg/connectivity/__init__.py
@@ -6,10 +6,11 @@
 from . import cross_frequency_coupling
 from .con_viz import (sensor_connectivity_3d, plot_grouped_connectivity_circle,
                       plot_generic_grouped_circle,
-                      plot_grouped_causality_circle)
+                      plot_grouped_causality_circle,
+                      plot_degree_circle, plot_lines_and_blobs,
+                      plot_labelled_group_connectivity_circle,
+                      plot_fica_grouped_circle)
 from .con_utils import (weighted_con_matrix, find_distances_matrix,
                         get_label_distances, make_annot_from_csv)
-from .con_circle import (plot_labelled_group_connectivity_circle,
-                         plot_fica_grouped_circle)
 from .causality import (do_mvar_evaluation, prepare_causality_matrix,
                         compute_order, make_frequency_bands)
